trying wasm on macosx
This will greatly improves overall build, because it takes ages to setup
LLVM 15 on ubuntu. That's why LLVM is superior to GCC. Sorry. I can't
use GCC on any platform with that level of simplicity - let alone
cross-compile using GCC (again, that easily). And MSVC is Windows-only -
so no real competition outside Windows. Also, not even C++ itself is
that portable TBH. Mostly because it depends on each compiler. Do I
expect the whole community to drop all compilers and go for LLVM only?
No. After all, no monopoly is good. We need competition to search for
new ideas in different places!